Renamed query managers.
