+++ src/api/ant-model-manager.ts
@@ -11,8 +11,8 @@ import { SingleResultQueryManager } from '../persistence/primary/query/SingleRes
 import { ApiModelManager, TAntQueryManager } from './api-model-manager';
 import { ApiModelConfig } from './config/api-model-config';
 import { ApiQueryConfig } from './config/api-query-config';
-import { AntMultipleResultQueryManager } from './query/AntMultipleResultQueryManager';
-import { AntSingleResultQueryManager } from './query/AntSingleResultQueryManager';
+import { AntMultipleResultQueryManager } from './query/ant-multiple-result-query-manager';
+import { AntSingleResultQueryManager } from './query/ant-single-result-query-manager';
 import { ApiQueryManager } from './query/api-query-manager';
 import { IAntMultipleResultQueryManager } from './query/IAntMultipleResultQueryManager';
 import { IAntSingleResultQueryManager } from './query/IAntSingleResultQueryManager';
