+++ src/persistence/primary/query/IPrimaryQueryManager.ts
@@ -1,8 +1,8 @@
 import { IEntity } from '../../../model/IEntity';
-import { IQueryManager } from './IQueryManager';
+import { QueryManager } from './query-manager';
 
 export interface IBasePrimaryQueryManager<TEntity extends IEntity, TResult extends TEntity | TEntity[]>
-  extends IQueryManager<TEntity, TResult> {
+  extends QueryManager<TEntity, TResult> {
   /**
    * True if the queries managed can return multiple results.
    */
